+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Test/interim releases are not documented here.
 - Transpiler - fix `FinC.ignore_unused(this)`
     - https://github.com/fin-language/fin/issues/45
 - Transpiler - fix interface generation for non-bool types
+- Transpiler - fix implicit conversion of classes to interfaces when returned from a method
 
 ### Changed
 - Transpiler - output `Source file` header info using Linux path separators